requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-cloudinary-storage==0.3.0
cloudinary==1.37.0
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple
dj-database-url
confection
psycopg==3.1.18
# GuardianUnivalle-Benito-Yucra
django-otp
django-two-factor-auth
pyotp==2.9.0
qrcode[pil]
Pillow
user-agents
